+++ librz/core/gadget_rop.c
@@ -1,0 +1,2 @@
+// SPDX-FileCopyrightText: 2026 MrQuantum1915 <darshanpatelgdh@gmail.com>
+// SPDX-License-Identifier: LGPL-3.0-only
