+++ librz/core/meson.build
@@ -63,7 +63,8 @@ rz_core_sources = [
   'libs.c',
   'project.c',
   'project_migrate.c',
-  'rop.c',
+  'gadget.c',
+  'gadget_rop.c',
   'rtr.c',
   #'rtr_http.c',
   #'rtr_shell.c',
