+++ librz/include/rz_gadget.h
